feat: Implement Phase 2 Procedural Loot System
I've implemented the procedural loot system for Phase 2. This builds upon the core gameplay loop from Phase 1.

Key features implemented:
- Weapon Data Structures: Created `Weapon.cs`, `WeaponPart.cs`, and enums to define the properties of all weapons. `WeaponPart` is a ScriptableObject to allow for easy creation of components in the editor.
- LootManager: A singleton that procedurally generates weapons by combining random parts from predefined pools.
- Loot Drops: Enemies now have a chance to drop a weapon upon death. A `WeaponPickup` object represents the dropped loot in the game world.
- Player Integration: The `PlayerController` has been updated to equip and use these procedurally generated weapons. The player's firing stats (damage, fire rate) are now driven by their equipped weapon.

+++ Assets/Scripts/Projectile.cs
@@ -6,7 +6,7 @@ public class Projectile : MonoBehaviour
 {
     [Header("Configuration")]
     public float speed = 20f;
-    public int damage = 1;
+    public float damage = 1f; // Damage is now a float to match the Weapon's stats
 
     private float topBound = 12f; // A simple boundary for when the projectile is considered off-screen
 
@@ -43,8 +43,9 @@ public class Projectile : MonoBehaviour
         EnemyAI enemy = other.GetComponent<EnemyAI>();
         if (enemy != null)
         {
-            // If it's an enemy, deal damage to it
-            enemy.TakeDamage(damage);
+            // If it's an enemy, deal damage to it.
+            // We cast the float damage to an int because the enemy's health is an integer.
+            enemy.TakeDamage((int)damage);
 
             // Destroy the projectile immediately after hitting an enemy
             Destroy(gameObject);
